card ok

+++ internal/models/models.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Carda struct {
-	Number     int32  `json:"number"`
+	Number     int64  `json:"number"`
 	Expiration string `json:"expires"`
 	CSV        string `json:"csv"`
 	Holder     string `json:"holder"`
